fix error creating fhd-skin

+++ usr/lib/enigma2/python/Plugins/Extensions/MyMetrixLite/MainSettingsView.py
@@ -1136,10 +1136,11 @@ class MainSettingsView(Screen):
                         move(dpath + file,dpath + file + ".hd")
                     copy(spath + file,dpath + file)
                 else:
-                    if not path.exists(dpath + file + ".del"):
-                        f = open(dpath + file + ".del", "w")
-                        f.close()
-                    copy(spath + file,dpath + file)
+                    if path.exists(dpath):
+                        if not path.exists(dpath + file + ".del"):
+                            f = open(dpath + file + ".del", "w")
+                            f.close()
+                        copy(spath + file,dpath + file)
 
         if target == "HD" and path.exists(dpath):
             for file in listdir(dpath):
